Fixed covid.csv final column, added plotly gif test
I was having issues with the bcr library not wanting to process our data after I fixed the final column of the dataset being combined (WIWY instead of WI,WY). Plus, the person who is running it needs to have FFMPEG installed, which is kinda a hassle on Windows. Therefore, I started to investiage the option of using plotly to export a gif or series of images, which we can iterate through in-application without loading the exported MP4 or gif to tkinter.

I think we might also have problems with stacking transparent images onto one another in Tkinter using the label element, as that doesn't seem to be supported without creating a combined image or using a canvas ( https://stackoverflow.com/a/49353569 and https://stackoverflow.com/a/56555164 and https://stackoverflow.com/questions/33904942/tkinter-image-transparency ).